refatorando o codigo

+++ src/pages/feed.js
@@ -8,19 +8,13 @@ const logout = () => {
 };
 
 const deletePost = () => {
+  const confirmDelete = confirm('Deseja mesmo deletar?')
+  if(confirmDelete){
   app.db.collection('posts').doc(id).delete().then(() => {
     document.getElementById(id).parentElement.parentElement.remove();
-  });
+  });}
 };
 
-const checkUser = (docUid) => {
-  const user = app.auth.currentUser.uid;
-  if (user !== docUid) {
-    document.getElementsByName(docUid).forEach((bt) => {
-      bt.remove();
-    });
-  }
-};
 
 const makePostEditable = () => { 
   document.getElementsByClassName(id)[0].firstChild.parentElement.contentEditable = true
@@ -33,30 +27,37 @@ const saveEditPost = () => {
   db.collection('posts').doc(id).update({text: pText.textContent, date: new Date().toLocaleString('pt-BR').slice(0, 16),})
 }
 
-const postTemplate = (doc) => {
-  document.querySelector('.posts').innerHTML
-    += `
-    <div class='posted container-post' data-id=${doc.id}> 
-      <p class='posted posted-name'> ${doc.data().name}
-      ${udButton({type: 'button', class: 'delete-btn', name: doc.data().user, id: doc.id, onClick: deletePost, title: 'X' })}
-      </p>
-      <p class='posted text ${doc.id}'> ${doc.data().text} |</p>
-      ${udButton({type: 'button', class: 'edit-btn', name: doc.data().user, id: doc.id, onClick: makePostEditable, title: 'Editar' })}
-      ${udButton({type: 'button', class: 'save-btn', name: doc.data().user, id: doc.id, onClick: saveEditPost, title: 'Salvar' })}
-      <p>${doc.data().date}</p>
-    </div>`;
-
-  checkUser(doc.data().user);
-  checkUser(doc.data().user);
+const checkUserEdit = (doc) => {
+  const user = app.auth.currentUser.uid;
+  if (user == doc.user) {
+    return `
+      ${udButton({type: 'button', class: 'edit-btn', name: doc.user, id: doc.id, onClick: makePostEditable, title: 'Editar' })}
+      ${udButton({type: 'button', class: 'save-btn', name: doc.user, id: doc.id, onClick: saveEditPost, title: 'Salvar' })}
+    `;
+  } 
+  return '';
 };
 
-const showPosts = () => {
-  app.db.collection('posts').orderBy('timestamp', 'desc').get()
-    .then((snapshot) => {
-      snapshot.docs.forEach((doc) => {
-        postTemplate(doc);
-      });
-    });
+const checkUserDelete =(doc) =>{
+  const user = app.auth.currentUser.uid;
+  if (user == doc.user) {
+    return `
+  ${udButton({type: 'button', class: 'delete-btn', name: doc.user, id: doc.id, onClick: deletePost, title: 'X' })}`
+
+}
+return '';
+};
+
+const postTemplate = (doc) => {
+  return `
+    <div class='posted container-post' data-id=${doc.id}> 
+      <p class='posted posted-name'> ${doc.name}
+      ${checkUserDelete(doc)}
+      </p>
+      <p class='posted text ${doc.id}'> ${doc.text} |</p>
+      ${checkUserEdit(doc)}
+      <p>${doc.date}</p>
+    </div>`;
 };
 
 const newPost = () => {
@@ -69,14 +70,28 @@ const newPost = () => {
     timestamp: new Date().getTime(),
     date: new Date().toLocaleString('pt-BR').slice(0, 16),
   };
-  app.db.collection('posts').add(post).then(() => {
-    postsSpace.innerHTML = '';
-    app.showPosts();
+  app.db.collection('posts').add(post).then((docRef) => {
+    const docPost = {
+      ...post,
+      id: docRef.id
+    }
+
+    document.querySelector('.posts').insertAdjacentHTML('afterbegin', app.postTemplate(docPost))
+
     textArea.value = '';
   });
 };
 
-function Feed() {
+const Feed= (props)=> {
+  let postsTemplate = '';
+  props.posts.forEach((post) => {
+    const docPost = {
+      ...post.data(),
+      id: post.id
+    }
+    postsTemplate += postTemplate(docPost)
+  });
+
   const template = `
     <section class="container">
       <section class="container">
@@ -84,7 +99,7 @@ function Feed() {
       ${Button({
     type: 'button', title: 'Postar', class: 'primary-button', onClick: newPost,
   })}
-        <div class="posts">  </div>
+        <div class="posts"> ${postsTemplate} </div>
       </section>
       ${Button({
     type: 'button', title: 'Logout', class: 'primary-button', onClick: logout,
@@ -95,11 +110,9 @@ function Feed() {
   return template;
 }
 
-window.onhashchange = showPosts;
-window.onload = showPosts;
 
 window.app = {
-  showPosts,
+  postTemplate,
   db: firebase.firestore(),
   auth: firebase.auth(),
 };
